Move schemas under V1 namespace
