+++ src/StoreApi/Schemas/V1/AbstractSchema.php
@@ -1,5 +1,5 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Schemas;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1;
 
 use Automattic\WooCommerce\Blocks\Domain\Services\ExtendRestApi;
