+++ src/StoreApi/Schemas/V1/CartShippingRateSchema.php
@@ -1,5 +1,5 @@
 <?php
-namespace Automattic\WooCommerce\Blocks\StoreApi\Schemas;
+namespace Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1;
 
 use WC_Shipping_Rate as ShippingRate;
