+++ src/deprecated.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * Class Aliases for graceful Backwards compatibility.
+ *
+ * This file is autoloaded via composer.json and maps the old namespaces to new namespaces.
+ */
+
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\AbstractAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\AbstractAddressSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\AbstractSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\AbstractSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\BillingAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\BillingAddressSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartCouponSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartCouponSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartExtensionsSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartExtensionsSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartFeeSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartFeeSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartItemSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartItemSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CartShippingRateSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CartShippingRateSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\CheckoutSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\CheckoutSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ErrorSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ErrorSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ImageAttachmentSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ImageAttachmentSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\OrderCouponSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\OrderCouponSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductAttributeSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductAttributeSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductCategorySchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductCategorySchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductCollectionDataSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductCollectionDataSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductReviewSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductReviewSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ProductSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ProductSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\ShippingAddressSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\ShippingAddressSchema::class );
+class_alias( Automattic\WooCommerce\Blocks\StoreApi\Schemas\V1\TermSchema::class, Automattic\WooCommerce\Blocks\StoreApi\Schemas\TermSchema::class );
